moves Simulation, SimulationMetadata from chatlab to simcore and adds reset_migration mgmt cmd
