+++ SimWorks/simai/async_client.py
@@ -12,7 +12,7 @@ from typing import Optional
 
 from asgiref.sync import sync_to_async
 from chatlab.models import Message
-from chatlab.models import Simulation
+from simcore.models import Simulation
 from . import prompts
 from .models import ResponseType
 from .output_schemas import message_schema, feedback_schema
